+++ RELEASE_NOTES.md
@@ -1,7 +1,9 @@
 Release Notes
 =============
-## vNext
+
+## 1.6.18
 * Resource Groups: Add support for multiple nested deployments targetting the same resource group
+* Resource Groups: Provide input parameters and key vault references to nested deployments.
 
 ## 1.6.17
 * Container Groups: Use an ARM expression to populate a secure environment variable.